+++ src/Umbraco.Web.UI.Client/package.json
@@ -62,6 +62,7 @@
 		"./member-group": "./dist-cms/packages/members/member-group/index.js",
 		"./member-type": "./dist-cms/packages/members/member-type/index.js",
 		"./member": "./dist-cms/packages/members/member/index.js",
+		"./member-public-access": "./dist-cms/packages/members/member-public-access/index.js",
 		"./menu": "./dist-cms/packages/core/menu/index.js",
 		"./modal": "./dist-cms/packages/core/modal/index.js",
 		"./models": "./dist-cms/packages/core/models/index.js",
