+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/entity-actions/constants.ts
@@ -4,5 +4,4 @@ export * from './culture-and-hostnames/constants.js';
 export * from './duplicate/constants.js';
 export * from './move-to/constants.js';
 export * from './notifications/constants.js';
-export * from './public-access/constants.js';
 export * from './sort-children-of/constants.js';
