+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/entity-actions/index.ts
@@ -1,3 +1,2 @@
-export * from './public-access/index.js';
 export * from './culture-and-hostnames/index.js';
 export * from './sort-children-of/index.js';
