+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/entity-actions/manifests.ts
@@ -8,7 +8,6 @@ import { manifests as createManifests } from './create/manifests.js';
 import { manifests as cultureAndHostnamesManifests } from './culture-and-hostnames/manifests.js';
 import { manifests as duplicateManifests } from './duplicate/manifests.js';
 import { manifests as moveManifests } from './move-to/manifests.js';
-import { manifests as publicAccessManifests } from './public-access/manifests.js';
 import { manifests as sortChildrenOfManifests } from './sort-children-of/manifests.js';
 import { manifests as notificationManifests } from './notifications/manifests.js';
 import { UMB_ENTITY_IS_TRASHED_CONDITION_ALIAS } from '@umbraco-cms/backoffice/recycle-bin';
@@ -67,7 +66,6 @@ export const manifests: Array<UmbExtensionManifest> = [
 	...cultureAndHostnamesManifests,
 	...duplicateManifests,
 	...moveManifests,
-	...publicAccessManifests,
 	...sortChildrenOfManifests,
 	...entityActions,
 	...notificationManifests,
