+++ src/Umbraco.Web.UI.Client/src/packages/members/manifests.ts
@@ -1,5 +1,6 @@
 import { manifests as memberGroupManifests } from './member-group/manifests.js';
 import { manifests as memberManifests } from './member/manifests.js';
+import { manifests as memberPublicAccessManifests } from './member-public-access/manifests.js';
 import { manifests as memberTypeManifests } from './member-type/manifests.js';
 import { manifests as sectionManifests } from './section/manifests.js';
 
@@ -11,6 +12,7 @@ import './member-group/components/index.js';
 export const manifests: Array<UmbExtensionManifest | UmbExtensionManifestKind> = [
 	...memberGroupManifests,
 	...memberManifests,
+	...memberPublicAccessManifests,
 	...memberTypeManifests,
 	...sectionManifests,
 ];
