+++ src/Umbraco.Web.UI.Client/src/packages/members/member-public-access/index.ts
