+++ src/Umbraco.Web.UI.Client/src/packages/members/member-public-access/manifests.ts
@@ -1,5 +1,5 @@
-import { UMB_DOCUMENT_ENTITY_TYPE } from '../../entity.js';
-import { UMB_USER_PERMISSION_DOCUMENT_PUBLIC_ACCESS } from '../../user-permissions/index.js';
+import { UMB_DOCUMENT_ENTITY_TYPE } from '../../documents/documents/entity.js';
+import { UMB_USER_PERMISSION_DOCUMENT_PUBLIC_ACCESS } from '../../documents/documents/user-permissions/index.js';
 import { UMB_SECTION_USER_PERMISSION_CONDITION_ALIAS } from '@umbraco-cms/backoffice/section';
 import { UMB_ENTITY_IS_NOT_TRASHED_CONDITION_ALIAS } from '@umbraco-cms/backoffice/recycle-bin';
