+++ src/Umbraco.Web.UI.Client/src/packages/members/member-public-access/repository/manifests.ts
