+++ src/Umbraco.Web.UI.Client/tsconfig.json
@@ -91,6 +91,7 @@ DON'T EDIT THIS FILE DIRECTLY. It is generated by /devops/tsconfig/index.js
 			"@umbraco-cms/backoffice/member-group": ["./src/packages/members/member-group/index.ts"],
 			"@umbraco-cms/backoffice/member-type": ["./src/packages/members/member-type/index.ts"],
 			"@umbraco-cms/backoffice/member": ["./src/packages/members/member/index.ts"],
+			"@umbraco-cms/backoffice/member-public-access": ["./src/packages/members/member-public-access/index.ts"],
 			"@umbraco-cms/backoffice/menu": ["./src/packages/core/menu/index.ts"],
 			"@umbraco-cms/backoffice/modal": ["./src/packages/core/modal/index.ts"],
 			"@umbraco-cms/backoffice/models": ["./src/packages/core/models/index.ts"],
